feat: add args for dfcache
