+++ dragonfly-client/src/bin/dfcache/stat.rs
@@ -22,7 +22,6 @@ use dragonfly_client_core::{
     Error, Result,
 };
 use humantime::format_duration;
-use std::path::Path;
 use std::time::Duration;
 use tabled::{
     settings::{object::Rows, Alignment, Modify, Style},
@@ -37,15 +36,66 @@ use super::*;
 pub struct StatCommand {
     #[arg(help = "Specify the persistent cache task ID to stat")]
     id: String,
+
+    #[arg(
+        short = 'e',
+        long = "endpoint",
+        default_value_os_t = dfdaemon::default_download_unix_socket_path(),
+        help = "Endpoint of dfdaemon's GRPC server"
+    )]
+    endpoint: PathBuf,
+
+    #[arg(
+        short = 'l',
+        long,
+        default_value = "info",
+        help = "Specify the logging level [trace, debug, info, warn, error]"
+    )]
+    log_level: Level,
+
+    #[arg(
+        long,
+        default_value_os_t = dfcache::default_dfcache_log_dir(),
+        help = "Specify the log directory"
+    )]
+    log_dir: PathBuf,
+
+    #[arg(
+        long,
+        default_value_t = 6,
+        help = "Specify the max number of log files"
+    )]
+    log_max_files: usize,
+
+    #[arg(
+        long = "verbose",
+        default_value_t = false,
+        help = "Specify whether to print log"
+    )]
+    verbose: bool,
 }
 
 /// Implement the execute for StatCommand.
 impl StatCommand {
     /// execute executes the stat command.
-    pub async fn execute(&self, endpoint: &Path) -> Result<()> {
+    pub async fn execute(&self) -> Result<()> {
+        // Parse command line arguments.
+        Args::parse();
+
+        // Initialize tracing.
+        let _guards = init_tracing(
+            dfcache::NAME,
+            self.log_dir.clone(),
+            self.log_level,
+            self.log_max_files,
+            None,
+            false,
+            self.verbose,
+        );
+
         // Get dfdaemon download client.
         let dfdaemon_download_client =
-            match get_dfdaemon_download_client(endpoint.to_path_buf()).await {
+            match get_dfdaemon_download_client(self.endpoint.to_path_buf()).await {
                 Ok(client) => client,
                 Err(err) => {
                     println!(
@@ -71,7 +121,7 @@ impl StatCommand {
                         style::Bold,
                         style::Reset,
                         err,
-                        endpoint.to_string_lossy(),
+                        self.endpoint.to_string_lossy(),
                     );
 
                     println!(
